System.Shell.QQ

Description

Quasiquoters for shell commands

Documentation

sh :: QuasiQuoter

QuasiQuoter for shell commands in default shell

Works only for expressions (obviously)

>>> return 3 :: IO [sh|blah|]

<interactive>:24:16:
    Exception when trying to run compile-time code:
      this quasiquoter does not support splicing types
      Code: quoteType sh "blah"

shell :: FilePath -> QuasiQuoter

class Eval r where

Different interesting return types for sh QuasiQuoter

Instances mostly resemble the types of things in System.Process

eval is not supposed to be used directly, user is expected to use sh quasiquoter instead

Methods

eval :: String -> [String] -> r

class Embed a where

Embed haskell values into shell scripts

Instances provided for all Prelude data types for which it makes sense

I recommend using -XExtendedDefaultRules for modules where you want to embed values, it would save for annoying type annotations for numeric literals

 embed . embed = embed

Methods

embed :: a -> String