System.Shell.QQ.ShellT

Description

Example DSL made on top of System.Shell.QQ

Provides [semi-]convenient way to run shell commands and parse their output.

DSL does not use any custom quasiquoters but provides Eval instances for custom datatypes that implement desired semantics

Documentation

>>> :set -XQuasiQuotes
>>> import System.Shell.QQ
>>> let lengths = [sh|while read line; do echo ${#line}; done|] :: String -> ShellT IO String

newtype ShellT m a

Shell commands sequencing result

Every shell command results either in failure (thus provides non-zero exit code and stderr) or some value (typically its stdout)

For example:

>>> runShellT $ [sh|echo -e "hello\nworld!!!"|] >>= lengths
Right "5\n8\n"

ShellT implements the usual Alternative instance semantics:

>>> runShellT $ [sh|exit 1|] <|> [sh|echo hello|]
Right "hello\n"

If everything fails, then last failure is returned:

>>> do Left (Last (Just (Failure _ i _))) <- runShellT $ [sh|exit 1|] <|> [sh|exit 3|]; print i
3

Constructors

ShellT

Fields

unShellT :: EitherT (Last Failure) m a

data Failure

Failed command with exit code and stderr

Constructors

Failure Command Int String

data Command

Command name and arguments

Constructors

Command String [String]

runShellT :: ShellT m a -> m (Either (Last Failure) a)

Run shell commands and get the result

(>>!) :: Monad m => ShellT m a -> (String -> ShellT m b) -> ShellT m b

Pass stderr of failed shell command to function

If nothing has failed, we do not have stderr, really:

>>> runShellT $ [sh|echo -e "hello\nworld!!!">&2|] >>! lengths
Left (Last {getLast = Nothing})

If something has failed, we do have stderr to play with:

>>> runShellT $ [sh|echo -e "hello\nworld!!!">&2; exit 1|] >>! lengths
Right "5\n8\n"

And playing may involve arbitrary Haskell functions, of course:

>>> runShellT $ [sh|echo -e "hello\nworld!!!">&2; exit 1|] >>! lengths . unlines . reverse . lines
Right "8\n5\n"